fsc header file created and minor code correction in pdb
